Refactor UI styling logic by introducing StyleManager for improved maintainability and consistency
